requirements.txt
torch >= 2.0
torchvision
numpy
pillow
scikit-learn
wandb


dataclasses
tqdm
